Move timeline columns
This aligns with how we structure our other resources where component helpers are in the "helpers" directory.
